Reorganize code
Prepare for a future PR where the macros crate needs to handle multiple different derives, not just one.
