+++ execution-plan-macros/tests/main.rs
@@ -1,39 +1,8 @@
 use kittycad_execution_plan_macros::ExecutionPlanValue;
 use kittycad_execution_plan_traits::{Primitive, Value};
 
-#[derive(ExecutionPlanValue)]
-struct FooConcrete {
-    f: f64,
-    i: usize,
-    o: Option<usize>,
-}
-
-#[derive(ExecutionPlanValue)]
-enum FooEnum {
-    A { x: usize, y: usize },
-    B { z: usize, w: usize },
-    C(usize, String, f64, f32),
-    D,
-}
-
-mod generics {
-    use kittycad_execution_plan_macros::ExecutionPlanValue;
-    use kittycad_execution_plan_traits::{Primitive, Value};
-
-    #[derive(ExecutionPlanValue)]
-    struct FooGenericWithDefault<T = f32>
-    where
-        Primitive: From<T>,
-        T: Value,
-    {
-        f: f64,
-        i: usize,
-        t: T,
-    }
-}
-
 #[test]
-fn test_derive_on_enum() {
+fn test_derive_value_on_enum() {
     #[derive(ExecutionPlanValue, Eq, PartialEq, Debug, Clone)]
     enum FooEnum {
         A { x: usize },
@@ -87,7 +56,7 @@ fn test_derive_on_enum() {
 }
 
 #[test]
-fn test_derive_on_struct() {
+fn test_derive_value_on_struct() {
     #[derive(ExecutionPlanValue, PartialEq, Debug, Clone)]
     struct MyStruct {
         f: f64,
